Create HomeViewModel & HomeView Refactor
