+++ FitnessApp/Home/Models/Activity.swift
@@ -1,0 +1,18 @@
+//
+//  Activity.swift
+//  FitnessApp
+//
+//  Created by Doğan Ensar Papuçcuoğlu on 3.01.2025.
+//
+
+import SwiftUI
+
+struct Activity {
+    let id: Int
+    let title: String
+    let subtitle: String
+    let image: String
+    let tintColor: Color
+    let amount: String
+}
+
