+++ FitnessApp/Home/Models/Workout.swift
@@ -1,0 +1,18 @@
+//
+//  Workout.swift
+//  FitnessApp
+//
+//  Created by Doğan Ensar Papuçcuoğlu on 3.01.2025.
+//
+
+import SwiftUI
+
+struct Workout: Identifiable {
+    let id: Int
+    let title: String
+    let image: String
+    let duration: String
+    let calories: String
+    let date: String
+    let tintColor: Color
+}
