+++ FitnessApp/Home/Views/HomeView.swift
@@ -8,24 +8,7 @@
 import SwiftUI
 
 struct HomeView: View {
-    @State var calories: Int = 123
-    @State var active: Int = 52
-    @State var stand: Int = 8
-    
-    var mockActivities: [Activity] = [
-        Activity(id: 0, title: "Today Steps", subtitle: "Goal 10,000", image: "figure.walk", tintColor: .green, amount: "6121"),
-        Activity(id: 1, title: "Today Steps", subtitle: "Goal 10,000", image: "figure.walk", tintColor: .red, amount: "6121"),
-        Activity(id: 2, title: "Today Steps", subtitle: "Goal 10,000", image: "figure.walk", tintColor: .blue, amount: "6121"),
-        Activity(id: 3, title: "Today Steps", subtitle: "Goal 10,000", image: "figure.walk", tintColor: .purple, amount: "6121")
-    ]
-    
-    var mockWorkouts: [Workout] = [
-        Workout(id: 0, title: "Running", image: "figure.run", duration: "23 mins", calories: "341 kcal", date: "Aug 3", tintColor: .green),
-        Workout(id: 1, title: "Running", image: "figure.run", duration: "23 mins", calories: "341 kcal", date: "Aug 3", tintColor: .red),
-        Workout(id: 2, title: "Running", image: "figure.run", duration: "23 mins", calories: "341 kcal", date: "Aug 3", tintColor: .blue),
-        Workout(id: 3, title: "Running", image: "figure.run", duration: "23 mins", calories: "341 kcal", date: "Aug 3", tintColor: .cyan)
-    ]
-    
+    @StateObject var viewModel = HomeViewModel()
     var body: some View {
         NavigationStack {
             ScrollView(showsIndicators: false) {
@@ -41,7 +24,7 @@ struct HomeView: View {
                                     .font(.callout)
                                     .bold()
                                     .foregroundStyle(.red)
-                                Text("123 kcal")
+                                Text("\(viewModel.calories) kcal")
                                     .bold()
                             }
                             .padding(.bottom)
@@ -50,7 +33,7 @@ struct HomeView: View {
                                     .font(.callout)
                                     .bold()
                                     .foregroundStyle(.green)
-                                Text("52 mins")
+                                Text("\(viewModel.active) mins")
                                     .bold()
                             }
                             .padding(.bottom)
@@ -59,16 +42,16 @@ struct HomeView: View {
                                     .font(.callout)
                                     .bold()
                                     .foregroundStyle(.blue)
-                                Text("8 hours")
+                                Text("\(viewModel.stand) hours")
                                     .bold()
                             }
                         }
                         Spacer()
                         ZStack {
-                            ProgressCircleView(progress: $calories, goal: 600, tint: .red)
-                            ProgressCircleView(progress: $active, goal: 60, tint: .green)
+                            ProgressCircleView(progress: $viewModel.calories, goal: 600, tint: .red)
+                            ProgressCircleView(progress: $viewModel.active, goal: 60, tint: .green)
                                 .padding(20)
-                            ProgressCircleView(progress: $stand, goal: 12, tint: .blue)
+                            ProgressCircleView(progress: $viewModel.stand, goal: 12, tint: .blue)
                                 .padding(40)
                         }
                         .padding(.horizontal)
@@ -91,7 +74,7 @@ struct HomeView: View {
                     }
                     .padding(.horizontal)
                     LazyVGrid(columns: Array(repeating: GridItem(spacing: 20), count: 2)) {
-                        ForEach(mockActivities, id: \.id) { activity in
+                        ForEach(viewModel.mockActivities, id: \.id) { activity in
                             ActivityCard(activity: activity)
                         }
                     }
@@ -113,7 +96,7 @@ struct HomeView: View {
                     .padding(.horizontal)
                     .padding(.top)
                     LazyVStack {
-                        ForEach(mockWorkouts, id: \.id) { workout in
+                        ForEach(viewModel.mockWorkouts, id: \.id) { workout in
                             WorkoutCard(workout: workout)
                         }
                     }
